Update Some Configs. Added bashrc_extension.
The new bashrc_extension is meant to be sourced by the `.bashrc` file
so that secrets can be kept in the `.bashrc` file and everything else
can be backed up openly via the `bashrc_extension` file. Added fuzzy
finding to lf and neovim, as well as directly on the command line via
the `f` command/function in bashrc_extension.