Improved .xinitrc logic. Added wallpaper functions.
Added a `wp` function to bashrc_extension for setting image file as the
wallpaper, and copying it to the `$HOME/.wallpaper` file for .xinitrc to
use when starting an X session. Also created a `wpf` function in bashrc_extensions
that uses fzf to let you recursively search through the directories of
`$HOME/Pictures/Wallpapers` for an image file to pass to `wp`.