`hist` changes + play function in bashrc_extension
The `hist` command now appends the current terminals history to the fzf,
and outputs your selection to stdout. The `ehist` command now does what
the old `hist` command did before this update, it runs `eval` on the
command you choose. Added a `play` function that uses `ytfzf` to
automatically play the audio for a whole page of results from your
search.