Updated play, wpf, and stty rows & cols.
Updated the play function to include a history that can be accessed
via fzf if you call it without passing any args. Updated the stty
rows and cols for my new window sizes. Updated wpf to show wallpapers
in a horizontal fzf preview window with img2txt if its installed.