nvim terminal, ollama, and qutebrowser
Added a 'term' alias for opening neovim with a terminal in the current
terminal session. Added ollama ask and ask_log functions for asking
specific models questions easily, output in bat for syntax highlighting
and saved to a temporary log file for reference. Added qutebrowser support
because I love the keyboard centric control it allows.